fix: always enable internal GraphQL handler
